Updates to `firedrake-parmmg` and testing due to Firedrake's new `pip install` (#114)
Firedrake is now installable via `pip`, with `firedrake-install` now
giving a `FutureWarning`. For an overview of changes on Firedrake side,
see https://github.com/firedrakeproject/firedrake/pull/4011. This PR
updates the `firedrake-parmmg` image and `reusable_test_suite.yml`
workflows to conform to those changes, as described below.

General:
- The `firedrake-parmmg` image is now only 2.4GB and the build time is
almost twice as fast (25 min faster)!
- Mesh adaptation repos are now cloned to the home directory (they used
to be under `/home/firedrake/firedrake/src`). Firedrake made this change
in their containers too so I am just following their structure.
- It is no longer necessary to create a virtual environment, so I don't
create one (to be consistent with Firedrake's containers again).

Dependencies:
- We used to install
[hypre](https://petsc.org/release/manualpages/PC/PCHYPRE/) (it's listed
in our
[petsc-configure-options.txt](https://github.com/mesh-adaptation/docs/blob/main/install/petsc_configure_options.txt))
but I don't think we need this so I do not install it here.
- Firedrake installs https://github.com/firedrakeproject/slepc in their
vanilla container, but we didn't have this in our old containers so I
don't install it now either.
- Firedrake no longer installs `vtk` by default, so I added it to
dependencies of packages.

Parallel tests:
- I realised that parallel tests did not contribute to the coverage
report. This is now fixed.
- The container now ships OpenMPI, so we can't run serial and parallel
tests with a single `pytest` call.
- Parallel tests are now run with `mpiexec -n N python3 -m pytest ...`,
which is strongly discouraged to be executed as root (see
https://docs.open-mpi.org/en/v5.0.x/man-openmpi/man1/mpirun.1.html#the-allow-run-as-root-option),
which we do in `reusable_test_suite.yml`. Long story short, I think this
is fine to do in our workflow since it uses a container so we're not
really concerned about something going wrong.
- The `reusable_test_suite.yml` now runs serial and parallel tests in
two separate steps, and the coverage report is generated in a further
separate step.

Related PRs:
- Animate: https://github.com/mesh-adaptation/animate/pull/175
- Goalie: https://github.com/mesh-adaptation/goalie/pull/291
- Movement: https://github.com/mesh-adaptation/movement/pull/159