Avoid workflow vulnerabilities (#119)
Fixes for issues identified in #117:
- template injection (see
https://woodruffw.github.io/zizmor/audits/#template-injection)
- excessive permissions (see
https://woodruffw.github.io/zizmor/audits/#excessive-permissions)

Edit: I completely removed the option to set the animate checkpoint
directory via input to the workflow. I don't think we would want to
change this from the default, so it was just overcomplicating things. So
if you agree with this change, the Animate test workflow should go from
this:
```
jobs:
  test_suite:
    uses: mesh-adaptation/docs/.github/workflows/reusable_test_suite.yml
    with:
      test-command: |
        export ANIMATE_CHECKPOINT_DIR=$(pwd)/.checkpoints
```
to this:
```
jobs:
  test_suite:
    uses: mesh-adaptation/docs/.github/workflows/reusable_test_suite.yml
```