Update docker images (#123)
There have been further changes to Firedrake's docker images. Previously
we used `firedrake-env` as the base image, where they would install some
system dependencies, set the `firedrake` user etc. But now they removed
this and they use the `root` user, with packages being installed in
`/opt`.

I think it's neater to keep it as it has been so far: be non-root and
have packages installed in `/home/firedrake`. So this PR keeps things as
they were: it just moves things from the old `firedrake-env` image to
`firedrake-parmmg` and tidies things up a bit.

The only change is that I install thetis, which I forgot to add in the
previous PR.

- I tested the `firedrake-parmmg` image with with Goalie in
https://github.com/mesh-adaptation/goalie/pull/299
- I tested the `firedrake-um2n` image locally and everything works well.
The CI here fails because it uses the `firedrake-parmmg:latest` image
which has still not been updated/pushed to ghcr.io. This will get
updated properly when pushed to main branch.