Update Firedrake installation instructions and workflow (#125)
Closes #112.

The Firedrake installation approach moved away from the
`firedrake-install` script and is now clearly separated in three
distinct steps that are described in
https://www.firedrakeproject.org/install.html and that are very easy for
users to follow. We follow their instructions almost exactly as
described, except that we pass some extra flags to PETSc's `configure`
(`--download-parmmg` etc.).

So here I removed a lot of our own installation instructions, as well as
related files that we used to aid the installation process.