Drop outdated Firedrake install test (#153)
As pointed out by @stephankramer, the `test_firedrake_install` CI
workflow is no longer needed because it doesn't actually test any
scripts that exist in the mesh adaptation repos and it's already covered
by the ParMmg Docker build workflow. Let's get rid of it because it just
keeps causing headaches.