Build firedrake-parmmg container straight from Dockerfile.vanilla (#155)
Instead of following firedrake master, as we used to, firedrake now has
a release and main version. This PR makes the CI follow firedrake main
(we can add release additionally TBD) and reconfigures the building of
the CI containers.

Rather than having our own base docker file with build instructions for
petsc and firedrake with our specific additional build flags to include
the (par)mmg libraries, this uses the generic
https://github.com/firedrakeproject/firedrake/blob/main/docker/Dockerfile.vanilla
which also allows for such additional depencies (additional petsc
configure flags) to be specified. This has the advantage of reducing our
maintenance because we don't have to keep adjusting whenever firedrake
changes their build instructions. It also will make it easier to
introduce a main and release version of the CI pipeline (if we decide to
go for that).

As part of this the containers (including the UM2N one) should now get
rebuild automatically in the right order whenever we change any of the
workflow or container files. This turned out to be a little trickier
than I thought: originally I tried to stick with separate workflows but
it turned out to be hard to get them to be executed in the right order -
in fact this already did not work correctly previously e.g. for
rebuilding firedrake-um2n *after* firedrake-parmmg (you can use
workflow_run/type:completed but if your combine it with other triggers
it still doesn't do what you want). So I decided to it to just combine
it in a single workflow with three jobs that reuse the
reusable_docker_build:
* build firedrake-parmmg-base that uses firedrake's Dockerfile.vanilla
with our desired build options
* build firedrake-parmmg
* build firedrake-um2n
where the last two builds are built on top of the previous, and
everything gets rebuilds whenever any of the dockerfile or workflow
files are changed. Next problem then is that even these jobs are
completely isolated, so they don't have immediate access to a container
built in a previous job. Now we can pull these of github's docker
repository, but the problem with that is that we only *push* after the
merge into main (for good reasons), so during a PR we can't actually
test things correctly. To deal with that the built container is now also
uploaded as an artifact that can be picked up in the next job. The
advantage of this is that you can now also download these artifacts
locally and test out the built docker images in a PR _before_ merging.

A small change in firedrake-parmmg is that the adaptivity packages are
now installed in /root rather than /home/firedrake - there no longer is
a user firedrake. Actually already in the current situation the
firedrake user is not actually functional: the user cannot use the
firedrake that was installed as root (already in the current setup). So
in the current CI we workaround this by logging into the containers as
root, meaning we don't actually test that things work if you use the
container for yourself and get logged in by default as user firedrake.

Note that merging this will break the CI in all packages as we haven't
rebuilt (been able to) for a long time so all packages are broken at the
moment with either release or main firedrake (see e.g.
https://github.com/mesh-adaptation/animate/pull/205). I propose we merge
this first, fix things to work with firedrake main and then discuss
whether we want to have our own release and main versions as well...

---------

Co-authored-by: Joe Wallwork <22053413+jwallwork23@users.noreply.github.com>