Install submodules (#165)
Supersedes #164.

In this PR, we instead set up submodules when installing the package in
the CI. If the package has no submodules then this should still work.
Once `adapt_common` is added as a submodule, that will then be part of
the install.

Also adding a `CODEOWNERS` file while I'm at it.

---------

Co-authored-by: Stephan Kramer <s.kramer@imperial.ac.uk>